# source/frontend/src/app/core/api/generated/openapi-core.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/usuarios": {
      "get": {
        "operationId": "UsersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Optional field name for sorting.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive text filter for representative read endpoints.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE",
                "LOCKED"
              ]
            }
          },
          {
            "name": "profileId",
            "required": false,
            "in": "query",
            "description": "Filter users by assigned profile id.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List admin users with current profile assignments."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a new admin user."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/admin/usuarios/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update admin user identity/status."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/admin/usuarios/{id}/perfis": {
      "put": {
        "operationId": "UsersController_assignProfiles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignProfilesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replace admin user profile assignments."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/admin/usuarios/{id}/papeis-diretos": {
      "put": {
        "operationId": "UsersController_assignDirectRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDirectRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replace direct role snapshots for a user."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/admin/parametros/sistema": {
      "get": {
        "operationId": "SystemParametersController_listSystem",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tenant system parameters."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "system-parameters"
        ]
      },
      "put": {
        "operationId": "SystemParametersController_upsertSystem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSystemParametersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upsert tenant system parameters."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "system-parameters"
        ]
      }
    },
    "/api/v1/admin/parametros/globais": {
      "get": {
        "operationId": "SystemParametersController_listGlobal",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Global parameters."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "system-parameters"
        ]
      }
    },
    "/api/v1/admin/parametros/globais/{chave}": {
      "put": {
        "operationId": "SystemParametersController_upsertGlobal",
        "parameters": [
          {
            "name": "chave",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertGlobalParameterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upsert one global parameter."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "system-parameters"
        ]
      }
    },
    "/api/v1/admin/feature-flags/{chave}": {
      "patch": {
        "operationId": "FeatureFlagsController_toggle",
        "parameters": [
          {
            "name": "chave",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleFeatureFlagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Toggle a tenant feature flag."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "feature-flags"
        ]
      }
    },
...
  },
  "info": {
    "title": "SGP API",
    "description": "Modern SGP NestJS API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "cognitoSub": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "LOCKED"
            ]
          }
        },
        "required": [
          "login",
          "name"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "LOCKED"
            ]
          }
        }
      },
      "AssignProfilesDto": {
        "type": "object",
        "properties": {
          "perfis": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "perfis"
        ]
      },
      "AssignDirectRolesDto": {
        "type": "object",
        "properties": {
          "papeis": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "papeis"
        ]
      },
...
      "UpsertSystemParametersDto": {
        "type": "object",
        "properties": {
          "values": {
            "type": "object",
            "description": "Map of system parameter keys and values persisted under the system namespace."
          }
        },
        "required": [
          "values"
        ]
      },
      "UpsertGlobalParameterDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "object",
            "description": "Canonical global parameter value payload."
          }
        }
      },
      "ToggleFeatureFlagDto": {
        "type": "object",
        "properties": {
          "ativo": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "ativo"
        ]
      },
...
    }
  }
}